% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukghg.R
\name{calcAlpha}
\alias{calcAlpha}
\title{Calculate values of alpha, the coefficient of variation in time}
\usage{
calcAlpha(
  ghgName = c("ch4", "co2", "n2o"),
  datect,
  sectorList = 1:10,
  timeScales = c(TRUE, TRUE, TRUE, TRUE),
  beta_df = data.frame(sector = 1:10, beta_year = rep(1, 10), beta_yday = rep(1, 10),
    beta_wday = rep(1, 10), beta_hour = rep(1, 10))
)
}
\arguments{
\item{ghgName}{Greenhouse gas: one of "ch4", "co2", or "n2o". Defaults to "ch4".}

\item{datect}{A vector of timestamps in POSIXct format.}

\item{sectorList}{A vector of sector numbers for which alpha values should be returned, e.g. c(1,3,7). Defaults to all.}

\item{timeScales}{A vector of logicals for including variation at inter-annual, seasonal, intra-weekly, and diurnal time scales (i.e. the POSIXlt variables year, yday, wday, and hour. Defaults to TRUE for all four.}

\item{beta_df}{A data frame of beta coefficients, which act as multipliers on each sector. Defaults to 1 for all, but This allows for parameter estimation by optimisation or MCMC.}
}
\description{
This function calculates values of alpha, the time coefficient.
}
\examples{
startDate <- as.POSIXct(strptime("01/06/2006", "\%d/\%m/\%Y"), tz = "UTC")
endDate   <- as.POSIXct(strptime("02/06/2006", "\%d/\%m/\%Y"), tz = "UTC")
# create a sequence of dates
nTimes <- 2
datect <- seq(startDate, endDate, length = nTimes)
alpha_df <- calcAlpha("ch4", datect)
}
\seealso{
\code{\link{calcFlux}} for the higher-level function which calls this.
}
\keyword{alpha}
\keyword{internal}
