% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukghg.R
\name{calcFlux_bio}
\alias{calcFlux_bio}
\title{A calcFlux_bio Function}
\usage{
calcFlux_bio(ghgName = c("ch4", "co2", "n2o"), datect = datect,
  proj = c("OSGB", "LonLat"), res = c("1", "20", "100"),
  unitType = c("mol", "g"), unitSIprefix = c("kilo", "none", "milli",
  "micro", "nano", "pico"), timeScales = c(TRUE, TRUE, TRUE, TRUE))
}
\arguments{
\item{ghgName}{Greenhouse gas: one of "ch4", "co2", or "n2o". Defaults to "ch4".}

\item{datect}{A vector of timestamps in POSIXct format.}

\item{proj}{Geographic projection for the gridded data, either "OSGB" or "LonLat".  Defaults to OSGB, LonLat not implemented yet.}

\item{res}{Resolution for the gridded data, either 1, 20 or 100 km.  Defaults to "1km". Not yet implemented for LonLat.}

\item{unitType}{Either molar ("mol") or mass-based ("g").}

\item{unitSIprefix}{Any standard SI prefix for the output units, from "kilo" to "pico".}

\item{timeScales}{A vector of logicals for including variation at inter-annual, seasonal, intra-weekly, and diurnal time scales (i.e. the POSIXlt variables year, yday, wday, and hour. Defaults to TRUE for all four.}
}
\description{
This function calculates biopogenic greenhouse gas fluxes from the UK, based on a spatio-temporal model and the national GHG inventory data.
}
\examples{
startDate <- as.POSIXct(strptime("01/06/2006", "\%d/\%m/\%Y"), tz = "UTC")
endDate   <- as.POSIXct(strptime("02/06/2006", "\%d/\%m/\%Y"), tz = "UTC")
# create a sequence of dates
nTimes <- 2
datect <- seq(startDate, endDate, length = nTimes)
myFlux <- calcFlux_bio("co2", datect, proj = "OSGB", res = "20", "mol", "micro")
myFlux <- calcFlux_bio("co2", datect + 6*3600, proj = "OSGB", res = "20", "mol", "micro")
myFlux <- calcFlux_bio("ch4", datect, "LonLat", "mol", "micro")
plot(datect, myFlux$total)
}
\seealso{
\code{\link{calcFlux}}, the higher-level function which calls this.
}
\keyword{flux}
\keyword{internal}
