% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukghg.R
\name{runShinyApp}
\alias{runShinyApp}
\title{Launches the shiny app for the ukghg package}
\usage{
runShinyApp()
}
\value{
shiny application object
}
\description{
This function provides a web browser interface to calculate greenhouse gas fluxes from the UK, based on a spatio-temporal model and the national GHG inventory data.
}
\keyword{app}
\keyword{shiny}
